chore: rename CorrelationContext to Baggage (#1687)
* chore: rename CorrelationContext to Baggage

* chore: lint
