+++ src/baggage/EntryValue.ts
