+++ src/index.ts
@@ -20,8 +20,8 @@ export * from './common/Time';
 export * from './context/context';
 export * from './context/propagation/TextMapPropagator';
 export * from './context/propagation/NoopTextMapPropagator';
-export * from './correlation_context/CorrelationContext';
-export * from './correlation_context/EntryValue';
+export * from './baggage/Baggage';
+export * from './baggage/EntryValue';
 export * from './metrics/BatchObserverResult';
 export * from './metrics/BoundInstrument';
 export * from './metrics/Meter';
